pyproject.toml
[tool.poetry]
name = "asdana"
version = "0.1.0"
description = "A powerful, highly customizable, and easy-to-use Discord bot."
authors = ["Ash Omaraie <ashkanomaraie@icloud.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
mypy = "^1.13.0"
python-dateutil = "^2.9.0"
psutil = "^6.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = { version = "^2.0.36", extras = ["asyncio"] }
asyncpg = "^0.30.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
black = "24.10.0"
google-api-python-client = "^2.151.0"
pytest = "*"
pytest-mock = "*"
pytest-asyncio = "*"
pylint = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "24.10.0"
tomli = "*"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
required-version = "24.10.0"
target-version = ["py310"]
include = '\.py$'

[tool.pytest.ini_options]
asyncio_mode = 'auto'